Move Response to a separate file
