+++ Sources/Get/Request.swift
@@ -80,31 +80,6 @@ public struct Request<Response> {
     }
 }
 
-/// A response with a value and associated metadata.
-public struct Response<T> {
-    public var value: T
-    /// Original response data.
-    public var data: Data
-    /// Original request.
-    public var request: URLRequest
-    public var response: URLResponse
-    public var statusCode: Int? { (response as? HTTPURLResponse)?.statusCode }
-    public var metrics: URLSessionTaskMetrics?
-
-    public init(value: T, data: Data, request: URLRequest, response: URLResponse, metrics: URLSessionTaskMetrics? = nil) {
-        self.value = value
-        self.data = data
-        self.request = request
-        self.response = response
-        self.metrics = metrics
-    }
-
-    func map<U>(_ closure: (T) -> U) -> Response<U> {
-        Response<U>(value: closure(value), data: data, request: request, response: response, metrics: metrics)
-    }
-}
-
 #if swift(>=5.6)
 extension Request: @unchecked Sendable {}
-extension Response: @unchecked Sendable where T: Sendable {}
 #endif
